+++ Changelog.md
@@ -65,6 +65,7 @@
 * Add parsing `Settings#default_tab_stop`
 * Add parsing `DocProperties`
 * Add parsing `Table#description` and `Table#title`
+* Add parsing `CommonNonVisualProperties#title` and `CommonNonVisualProperties#description`
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit